feat: implement signature request controller
Without this patch the user has no way of canceling signature requests.

This patch introduces a status field. It has 3 possible values: PENDING,
CANCELED and EXECUTED.
The controller allows canceling the signature request only when it's in
status PENDING and the caller authenticates as one of the owners of the
Safe.