feat: add collections entity
Without this patch the user of our GraphQL endpoints needs to go via
hyperboards to access the collections inside. Another drawback is that
via this query they can only access the hypercert id, so they'll have to
fire a separate query to combine the data.

This patch adds collections as a top level entity and exposes full
hypercert objects as child entities. This also applies to hyperboards as
the standalone collections entity is being reused there too.